implemented sim_world_add_particle and graavity application

+++ include/simulator.h
@@ -21,7 +21,7 @@ typedef struct {
 	vector accel;
 	float radius;
 	float mass;
-	size_t color;
+	uint32_t color;
 } Particle;
 
 /*
